added image
I have added the image in the assets folder. I removed the URL of image and instead added the path to downloaded one.
In previous commits I had added 3 input slider widgets, a dataframe of random generated numbers, and a line chart visualizing the data points. I also added a markdown text for the "About" page in the dashboard.